+++ tests/codegen/struct_nested_struct_anon.cpp
@@ -74,7 +74,7 @@ entry:
   store i32 %3, i32* %2, align 8
   %4 = bitcast i32* %"Foo::(anonymous at definitions.h:1:14).x" to i8*
   call void @llvm.lifetime.start.p0i8(i64 -1, i8* nonnull %4)
-  %probe_read = call i64 inttoptr (i64 4 to i64 (i8*, i64, i8*)*)(i32* nonnull %"Foo::(anonymous at definitions.h:1:14).x", i64 4, i8* nonnull %1)
+  %probe_read = call i64 inttoptr (i64 4 to i64 (i32*, i64, i8*)*)(i32* nonnull %"Foo::(anonymous at definitions.h:1:14).x", i64 4, i8* nonnull %1)
   %5 = load i32, i32* %"Foo::(anonymous at definitions.h:1:14).x", align 4
   call void @llvm.lifetime.end.p0i8(i64 -1, i8* nonnull %4)
   %6 = bitcast i64* %"@x_key" to i8*
@@ -85,7 +85,7 @@ entry:
   call void @llvm.lifetime.start.p0i8(i64 -1, i8* nonnull %8)
   store i64 %7, i64* %"@x_val", align 8
   %pseudo = call i64 @llvm.bpf.pseudo(i64 1, i64 1)
-  %update_elem = call i64 inttoptr (i64 2 to i64 (i8*, i8*, i8*, i64)*)(i64 %pseudo, i64* nonnull %"@x_key", i64* nonnull %"@x_val", i64 0)
+  %update_elem = call i64 inttoptr (i64 2 to i64 (i8*, i64*, i64*, i64)*)(i64 %pseudo, i64* nonnull %"@x_key", i64* nonnull %"@x_val", i64 0)
   call void @llvm.lifetime.end.p0i8(i64 -1, i8* nonnull %6)
   call void @llvm.lifetime.end.p0i8(i64 -1, i8* nonnull %8)
   ret i64 0
