+++ tests/codegen/struct_string_ptr.cpp
@@ -77,7 +77,7 @@ entry:
   call void @llvm.memset.p0i8.i64(i8* nonnull %3, i8 0, i64 64, i32 1, i1 false)
   %4 = bitcast i64* %Foo.str to i8*
   call void @llvm.lifetime.start.p0i8(i64 -1, i8* nonnull %4)
-  %probe_read = call i64 inttoptr (i64 4 to i64 (i8*, i64, i8*)*)(i64* nonnull %Foo.str, i64 8, i8* nonnull %1)
+  %probe_read = call i64 inttoptr (i64 4 to i64 (i64*, i64, i8*)*)(i64* nonnull %Foo.str, i64 8, i8* nonnull %1)
   %5 = load i64, i64* %Foo.str, align 8
   call void @llvm.lifetime.end.p0i8(i64 -1, i8* nonnull %4)
   %probe_read_str = call i64 inttoptr (i64 45 to i64 (i8*, i64, i8*)*)([64 x i8]* nonnull %str, i64 64, i64 %5)
@@ -167,7 +167,7 @@ entry:
   call void @llvm.memset.p0i8.i64(i8* nonnull %1, i8 0, i64 64, i32 1, i1 false)
   %2 = bitcast i64* %Foo.str to i8*
   call void @llvm.lifetime.start.p0i8(i64 -1, i8* nonnull %2)
-  %probe_read = call i64 inttoptr (i64 4 to i64 (i8*, i64, i8*)*)(i64* nonnull %Foo.str, i64 8, i8* null)
+  %probe_read = call i64 inttoptr (i64 4 to i64 (i64*, i64, i8*)*)(i64* nonnull %Foo.str, i64 8, i8* null)
   %3 = load i64, i64* %Foo.str, align 8
   call void @llvm.lifetime.end.p0i8(i64 -1, i8* nonnull %2)
   %probe_read_str = call i64 inttoptr (i64 45 to i64 (i8*, i64, i8*)*)([64 x i8]* nonnull %str, i64 64, i64 %3)
